WIP: Integrate ThunkState (see comments)
This integrates thunk state into the UI.

The next step would be to have a function that, given a BChan and a
Present to evaluate and a Path that it belongs to, will run an IO action
to evaluate that thing.

* Post to the BChan that evaluation is starting, with an async to
  support cancellation.
* Attempt evaluation.
* On success, post the Whnf to the BChan at the given Path.

The event loop will integrate the Whnf and Path into its Map in State.

Now instead of [ready] it will show the actual Whnf.

To show expanded data structures, one merely kicks off N concurrent eval
jobs to progress parts of the UI asynchronously, which, when pending,
cancelled or succeeded, will update the UI.